Jigsaw-proof ProcessorMetrics
Ensures only public, exported classes are used in gathering the CPU metrics. This future-proofs the code for post-Jigsaw versions of Java. It also means we need to be explicit about the JVM implementations we support (HotSpot and J9, for now) so that we use the exported interface, rather than rely on the inaccessible implementation class.