Make tests pass on JDK 9
Due to some Java EE modules not being resolved in the JDK anymore since Java 9 (see [0]), these dependencies need to be made explicit where needed at runtime.

Since micrometer-spring-legacy is not trying to be Java 9 compatible at runtime (since Spring Boot 1.5 does not support Java 9), these are only needed for the purpose of running the tests on Java 9 and later.

If the micrometer-jersey2 module is intended to be compatible with Java 9+, the currently test-scoped dependency may need to be included at runtime. This may be an issue with Jersey's Java 9 runtime compatible itself, also.

[0] https://docs.oracle.com/javase/9/migrate/#GUID-F640FA9D-FB66-4D85-AD2B-D931174C09A3