Simplify computation of the status code
Given that the status code is computed for 'RequestEvent' of type
'FINISHED', we can simply use the 'ContainerResponse' to figure out the
status code.
That way we handle both 'WebApplicationException' and
'MappableException' with registered 'ExceptionMapper'.

In cases in which the 'ContainerResponse' is not there, we fallback to
'Response.Status.INTERNAL_SERVER_ERROR.getStatusCode()'.