Elastic reporter improvements
This commit improves the elastic reporter behaviour in a few ways:

First, it uses way less objects by removing the builder class and the
usage of streams. Second it does not generate one big in memory string
of the metrics, but rather writes the metrics directly to the output
stream - there is still a string builder per metric.

The first line of the bulk requests now does not include the index name
and the type as those can be specified in the URL, thus reducing the
size of the bulk request.

The encoding of the basic auth data only needs to be done once in the
constructor instead of doing it in each request.

Lastly tests for each different kind of metric have been added.