Polish #1147
Rename Tuple → DynatraceBatchedPayload. The class is only needed for this one specific purpose and is package-private to the dynatrace package, so this avoids making it too generic. This also improves the readability by using meaningful field names.

Adds tests for cases that were not covered previously, and updates the logic to handle these cases.