Fix kafka consumer fetch manager metrics to be reported per topic and partition (#878)
The kafka consumer fetch manager metrics have been registered too many times and therewith
also for the wrong mbeans. Hence, when using prometheus as a registry, for example, only
the metrics registered for the wrong mbean were shown with NaN as values. With this fix
the metrics are registered for the right mbeans and in case they are provided for different levels
like per consumer and topic as well as per consumer, topic and partition they are only registered
once for the lowest level of details. Aggregations can be done with the monitoring tools.