Additional Datadog tests
An issue was reported about units conversion for Datadog, which has a strict list of units it accepts. While we had the logic to handle this, we did not have any tests for it. This adds testing and hopefully fixes an existing but disabled test so it can be run on the CI server.

See #1181