Microsoft Azure Monitoring needs JAXB at runtime
The tests for it were not being run before even though the code was there and seemingly configured. That was fixed by 2b73f064c8010047f0b4f57cd3669592962e6a4e, which then revealed that when running (tests) on Java 9+ JAXB would be needed. This adds it as a test runtime dependency and updates lockfiles accordingly.