Simplify build configuration for JDK versions > 8
The build continues to work with JDK 8 and 11, but Java 8 compatible artifacts can be reliably produced with just a JDK 11 build using the new Java compiler arg `--release`. This opens up the opportunity for us to increase our build's JDK version while maintaining compatibility with Java 8 and not increasing complexity. If we want to use Java 9+ language features in test code, that is also a possibility now, but will require all contributors to have at least that version of the JDK installed to build locally.