Don't run tests again when deploying snapshots
Our CI job is setup to run tests first on all changes. As a separate step, if the change is on a release branch, we will call the deploy script, which used to run tests again when publishing snapshots. Skipping tests when publishing snapshots should reduce the time required by almost half.