Added instrumentation for Apache HttpComponents HttpRequestExecutor
The `MicrometerHttpRequestExecutor` can be used to monitor the execution time
of all requests going through a `HttpClient`. It collects metrics by HTTP
method and HTTP status.

In addition, it allows to tag the metrics by remote scheme, host and port and
also to tag the requested URLs (or URL patterns).