Run release script and maven central sync from CI for release tags (#1245)
We had the build configuration and scripts available to do this already but the CI config was not setup for it. This makes it so maintainers do not have to run the release process locally and can instead make a release/tag in GitHub, which will start the new workflow. There is a manual approval stage between publishing to Jcenter and syncing to Maven Central so release artifacts can be consumed before final sync.