Increase no output timeout for Maven Central sync to 30 minutes
During the `mavenCentralSync` task, output is produced per artifact synced. It is conceivable that a single artifact will take more than 10 minutes to sync, and the default `no_output_timeout` is 10 minutes. This increases it to 30 minutes. See https://support.circleci.com/hc/en-us/articles/360007188574-Build-has-hit-timeout-limit
We haven't hit this issue yet because the underlying client making the sync call has a lower (3 minute) read timeout currently hard-coded (see spring-gradle-plugins/spring-bintray-plugin#3). Once this is increased, we want to make sure CircleCI's `no_output_timeout` is not the next cause of the sync failing.