Catch and log uncaught exceptions when calling PushMeterRegistry#publish
Scheduled tasks that throw uncaught exceptions stop being scheduled by default and the exception will not be logged. This made it impossible for users to report the cause of failures without custom code/debugging, and it is even difficult to notice when this happens due to the lack of logs. Now calls to the `publish` method are guarded by a try-catch that logs the exception. Implementations of `publish` should still implement proper exception handling, but this acts as a final safety net for bugs in those implementations.