Disambiguate Tag class to fix compilation error
Amazon's model package and the micrometer core instrument package are both being wildcard imported and both contain a `Tag` class, which was causing compilation failures due to the ambiguous reference to the `Tag` class in the code. The `Tag` class seems to have been added to Cloud Watch's package in a recent version, which is why this failure only recently appeared and only on master.

Resolves #1348