Batching logic was not following New Relic expectations (#1376)
A pack of 3000 events was broken into 1000 and 2000, which led to all events from the second part being thrown out. Large packs should be broken into batches of `batchSize` each, eg. 1000, 1000, 1000 in this example.

A log was also added which states the intent to send a certain number of
events. Because without it it took us too long to identify the problem.