Register DiskSpaceMetrics with a strong reference
Without a strong reference to the `DiskSpaceMetrics` instance, it and its reference to the File `path` would be garbage collected, resulting in the gauges reporting NaN for the value.
Fun fact: NaN is greater than zero. So tests were updated to check for this and test the garbage collection scenario specifically.

Fixes #1409