Skip bintrayPublish until all modules are uploaded to Bintray
The implementation of the bintrayPublish task causes it to be skipped if any files are published for the package and version. So running it after uploading each modules causes later modules to not have their files published. Instead, run the bintrayPublish task once after all modules have been uploaded. With all modules published to the same package and version, publication only needs to happen once for the whole package/version.

Resolves #1579