Downgrade reactor-netty / netty / reactor to restore prior behavior
This is a temporary solution until we have a long-term solution for the issue faced. With downgrading, the `sendOptions` needed to be added back on the UdpClient and TcpClient. We will also need to revert that when we eventually upgrade again. Since these dependencies are shaded, the downgrading is not expected to otherwise affect users.

Resolves #1684