Remove Kafka meters on unregistration notification (#1745)
Whenever a Kafka consumer is registered, metrics for it are also registered. However, when a Kafka consumer was closed, its metrics were not being removed. In use cases where Kafka consumers are continually created temporarily, this lead to ever increasing memory to hold all of the metrics for the already closed Kafka consumers.

The changes here remove corresponding metrics when Kafka consumers are closed.

Closes gh-1743