Rewrites the meter registration benchmark
This styles after some benchmarks in zipkin, where we typically look at
the p0.999 and gc.alloc.rate.norm to see if changes are notable or not.
This approach could certainly become more refined also :P

Notably, this changes the benchmarks to progressively show impact of
registration. Meanwhile, the registry is held in state hoping to avoid
surprise of OOM by ensuring all bench executions add another counter to
it.