#1796 StepTimer#max should always return max of last period.
This also includes a refactoring of the 'rolling' logic for step values.

Currently, this is copied between the classes.  I had to create a custom
Step value for a specific registry and some of the grief my team members
gave me was how much code I had to copy from the open source repo.  This
allows for customization with much less copying.  If you don't want
that, I will take it out.