Track client writes for UDP to start server after all writes
Also reduces the increments while the server is down to hopefully prevent a flush happening before all increments happen. A flush results in an exception while the server is down. The handler writing metrics will re-subscribe but the time in between the error and re-subscribe results in an unspecified number of metrics being dropped, which makes it hard to write a stable test.