Prevent negative count infinity bucket with Stackdriver distributions
Due to a lack of guarantee that histogram snapshots are self-consistent (see #998), the count value could be lower than the sum of the bucket counts. Before this fix the infinity bucket count would have a negative value, which Stackdriver rejects with an error. This uses zero (0) instead in this situation. This does not fix the potential for inconsistent values in histogram snapshots.

Resolves #1325