Implement callEnd method in OkHttpMetricsEventListener (#1938)
Calls could be left in the `callState` map previously as there was no applicable callback that removed them for cached responses. By implementing the `callEnd` callback, these cached response `Call`s are cleaned up and do not leak.

This also upgrades okhttp to 3.14.7 due to a bug in the version of okhttp we were using where callbacks did not happen for cached responses at all - making the test for this fix fail. Users will need to be on at least okhttp 3.14.2 for the fix to work.

Fixes #1937

Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>