Polish WavefrontSender changes
The `WavefrontSender` handles batching internally, so this removes partitioning we were doing on publishing meters. Renames private methods to reflect their modified behavior. Changes signature of `publishMetric` and `publishDistribution` to use the registry's `WavefrontSender`. Lowers log level of failed publish to WARN, which still may be too verbose on a per-metric/per-distribution basis to log publish failures. We can revisit this more in follow-ups.

See #1901