Handle common tags in KafkaMetrics
When comparing metrics for a meter with the same name but less tags, we need to consider any common tags that are configured in that comparison. This uses a hack to get the common tags when binding to a registry because we lack an API currently to retrieve common tags.