Fix KafkaMetrics to handle multiple clients and common tags (#1978)
Only check meters from the same client when checking if a meter with less tags already exists. Otherwise, meters from one client only remain.

* Handle common tags in KafkaMetrics

When comparing metrics for a meter with the same name but less tags, we need to consider any common tags that are configured in that comparison. This uses a hack to get the common tags when binding to a registry because we lack an API currently to retrieve common tags.

Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>

Fixes #1968