Do not remap id in remove method
Registering a meter to a `MeterRegistry` returns a meter with a mapped id (`MeterFilter`s `map` method applied to it), so we should allow users to pass that to the remove method and it work. The previous behavior mapped the ID (again), which was problematic if any `MeterFilter` mapping was not idempotent.

Fixes #1850