Remove redundant call to stop on close of GraphiteMeterRegistry
In the latest patch versions of Dropwizard's metrics-core library (version 4.1.18 and later), a call to `stop` on the reporter does a final publish, which results in an extra publish happening because the `close` method will also `stop` the reporter. This change effectively reduces the publishes on close from 3 to 2 if using metrics-core 4.1.18 or later. We don't want to eliminate the explicit call to `report` as that will result in no final publish for those using metrics-core 4.1.17 or earlier.

Partially resolves #2156