Remove redundant call to report on close of GraphiteMeterRegistry
In the latest patch versions of Dropwizard's metrics-core library (version 4.1.18 and later), calls to `stop` on the reporter does a final publish and calls to `close` will `stop` the reporter. Therefore, to have the final publish on close behavior we want, we only need to call `close` on the reporter. If a version earlier than 4.1.18 of metrics-core is used, there will no longer be a final publish on close - users are encouraged to use the latest patch version of metrics-core to maintain that behavior.

Resolves #2164