Re-include units in Datadog metadata
When we added units to metric payload in #1864, we also removed it from
the metadata payload, only making the query when the description was
set. The problem is that the setting the descroption without the other
fields actually override them, which means it broke when descriptions
were enabled.

This fixes it by still checking if description is set to keep the
performance enhancement, but set all the fields in the metadata.

Closes #1977