Avoid async log4j2 log events to be counted multiple times (#2183)
For async loggers, the metrics filter will be called multiple times, but we should only increment on the last call. This uses the `isEndOfBatch` method on the log event to determine this.

This also includes some changes to prevent flakiness in the test. The LoggerContext is cached in global state and should be cleared so each test runs in isolation.

Co-authored-by: Dennys Fredericci <dennys.fredericci@kbc.be>
Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>