Avoid StatsD connection infinite reconnect loop (#2225)
The `doOnDisconnected` callback is sometimes called twice for the same connection when an error causes the disconnect. This would result in an infinite reconnect loop with the logic to make a new connection and close the last connection. This tracks if the reconnect logic has happened for a connection already or not so it is only called once. It also removes the logic to close the previous connection (added for #2138) as there should not be a case when a prior connection hasn't already been closed. The connection leak should be fixed by not calling the reconnect logic twice for a disconnect.

Resolves #2177