Fixing backward compatibility of GAUGE->CUMULATIVE change
When metric was created using Micrometer version prior this change, it wouldn't be able to record new TimeSeries to the metric. The reason is that both MetricDescriptor and TimeSeries request has MetricKind and they need to be the same. As a backward compatible was of handling it, it will use MetricKind value that is already configured in Stackdriver and suggest deleting the metric and letting Micrometer recreate it with the new MetricKind.