Support metrics for multiple Tomcat MBeans of same type (#2262)
This works with multiple MBeans for any of the metric types now, which would be expected when there are multiple servlets registered, for example. To support the case when TomcatMetrics is bound before Tomcat MBeans are available, the NotificationListener is no longer unregistered when multiple MBeans could be returned in the query since we have no way to know when all have been registered. All NotificationListeners are still unregistered when `close` is called.

Resolves #1989