Fix logic to remove Kafka meters with less tags (#2312)
This lead to many exceptions when using Prometheus with multiple Kafka client IDs and some meters with topic tags never being registered. No longer considers client-id specially because all meters with less tags need to be removed before the new meter with more tags can be added. A test case is added with PrometheusMeterRegistry behavior and multiple client IDs with an increase in tags present.

Resolves #2212