Upgrade to Reactor 3.4 and Reactor Netty 1.0
This also switches to using the Reactor BOM for dependency management. Some internal changes to the StatsD registry usage of Reactor were necessary with the move from Processors to Sinks. The many, multicast, directBestEffort Sink is the closest match to DirectProcessor that we were using before.
With 1.0, reactor-netty now offers a core dependency that doesn't include HTTP, which we can use to save some bytes in the statsd JAR.

Resolves gh-2319