Note about wrapped ExecutorService types
It can be unclear why metrics are missing when passing a wrapped type, as the API seems to indicate any `Executor` or `ExecutorService` can be monitored. This makes explicit the types we can monitor and notes unwrapping should be done by the user, since we cannot know wrapper types from external code.

See gh-2443 and gh-2426