Clarify internal naming for JvmGcMetrics
In light of supporting additional garbage collectors, it's come up that some of the existing internal naming is not as specific as it could be. While there is often more than one memory pool belonging to the young generation in garbage collectors, we are interested in the one used for allocation for our metrics. The logic remains the same as before, but the naming is more specific.