Refactor JvmGcMetrics with less specialized logic for non-gen collectors
When we introduced support for non-generational collectors (Shenandoah, ZGC), the logic for their metrics was special case in the code. This refactor makes more code shared regardless of whether the collector is generational or not. Really, non-generational collectors do not have a promoted counter since there is no old generation. This refactoring should make it easier to support additional collectors (e.g. OpenJ9 collectors).