Upgrade to JMH plugin 0.5.3, latest JMH versions, write locks
There were some lock files missing for some scopes, and so ehcache was being resolved with the latest version that has a transitive dependency on ehcache-parent 2.25 which is not available in central. This caused the build to fail. That version is now explicitly set to the version used in the lock, as well as writing the missing lock files. This cleans up the benchmarks project a bit: removes the dependency on the samples project (unused), and updates to the latest versions of JMH dependencies.

Closes gh-2576