Rename netty native files so they can be loaded (#2611)
Our shaded version of Netty needs the native library files to be named according to the shaded package for them to be loaded. This change ensures that naming is applied when shading. This was needed for loading the MacOS native DNS library, without which errors happened. It will also allow the native epoll library to be loaded on Linux and therefore epoll will be used when it can be, which is different behavior than we've had since we excluded the epoll native library when shading.