Add reflection config for native images
Without this configuration, ProcessorMetrics and FileDescriptorMetrics are not available in a Micrometer application run as a native image. This explicitly lets the native image compiler know what class and methods will have reflection used on them. With this configuration, the ProcessorMetrics and FileDescriptorMetrics are the same in applications run on the JVM or as native images.