Remove reflection config for JDK 14+ getCpuLoad method
Since GraalVM currently doesn't offer a non-experimental release for a Java version 14 or above, we need not target support for it yet in native images. Once a stable GraalVM version for Java 14+ (presumably Java 17) is released, we may need to figure out how to support the Java 11 release and Java 17 release. It was the intention of our reflection config before this change to support both, but it does not work with Java 11 without the `--allow-incomplete-classpath` option.

Resolves gh-2670