Migrating to one lockfile per project instead of one per project per configuration
With Gradle 7, this is the default behavior (it is easier to track and handle).
This commit does not change the locked version of any of the dependencies.

You can reproduce this change the following way:
Executing ./gradlew sometask --update-locks somedependency updates the versions of the dependencies of the configurations that are involved.
Using a task that resolves all the configurations and providing a non-existent dependency, you should be able to migrate the lockfiles without updating the versions, e.g.:
./gradlew downloadDependencies --update-locks 'org.something.that:does-not-exisit-54321'

After this, manually checking the old files is required, e.g.: compileOnly files could left there that you need to delete.

fixes gh-2695