Fix TimeWindowMax for case of long period of inactivity (#2648)
Existing logic assumed that the rotate method would be called at least once per rotation interval * ring buffer length. This is true in many cases, but in the case of a stopped registry or a pull-based registry not being scraped for some time, the assumption may not hold true. Logic to handle this missing edge case has been added along with a test to check max is as expected.

Resolves gh-2647