Deprecate JsonUtils and remove implementation
The code for the prettyPrint method comes from Stackoverflow, which has a license incompatible with our Apache licensed code. The Stackoverflow code has been removed and the method/class deprecated for removal. On the off chance this API is used, I did not delete the whole class in a patch release. Though it was always intended to be an internal class, it was not clearly marked as such.

The effect of this is that the `HttpSender.Request` class' `toString` will no longer be pretty-printed for JSON requests. This is an acceptable compromise for now rather than coming up with an original implementation for non-critical functionality.

See gh-2762