Migrate Jersey instrumentation into micrometer-core (#2780)
* Migrate jersey code into core and deprecate micrometer-jersey2

The Jersey instrumentation has been updated to support both Jersey 2 and Jersey 3. It is confusing to have this in a micrometer-jersey2 module, and it would be more in line with our other instrumentation if it lived in micrometer-core.

* Update Jersey 3 sample to use core

Use the Jersey code in core rather than the deprecated micrometer-jersey2 module.