Initial timer listener proof-of-concept
Adapts the `Timer.Sample` to make the callbacks on the `TimerRecordingListener` for `onStart`, `onError`, `onStop`. Being the initial pass, this tries to keep things as minimal as possible. Even the `onError` callback could probably have been initially omitted since we should work on making error tagging more of a first-class citizen in Micrometer. Then using an `onError` callback will be more natural.