Deprecate ReactorNettySender
Having code in micrometer-core that depends on reactor-netty creates a dependency cycle between the two projects. In an effort to work towards eliminating that cycle, we are proactively deprecating the ReactorNettySender. If we receive feedback that users have use cases for it, we can consider moving it to the reactor-netty project or elsewhere. Lacking enough feedback, we will plan to remove this in the next feature release.

See gh-2802