`DefaultLongTaskTimer` histogram bucket counting logic fix (#2851)
* DefaultLongTaskTimer histogram bucket counting logic fix

The histogram was not correct in all cases before. The added test demonstrates a case that would fail with the previous logic. Buckets that should have been zero were not.

* Store all task durations for more consistency

We don't want to compare the current bucket to the current duration where the duration is potentially different on each iteration. We get all the durations first and then iterate, which is closer to the expected behavior of a true snapshot.

* Move test to DefaultLongTaskTimerTest

It's testing logic in the DefaultLongTaskTimer class, so it is more appropriate for it to be in that test class.