Add tests for current sample API and experiment
The tests demonstrate some scenarios and check the expected behavior.
An experiment with changing to `InheritableThreadLocal` is included. The commented code in `TimedRunnable` is not needed thanks to `InheritableThreadLocal`, but if uncommented that code can make things work with a regular `ThreadLocal`.
The `startedAndStoppedOnDifferentThread` test does not pass either way.