`jvm.gc.live.data.size` and `max` not updating for optavgpause/optthruput collectors (#2874)
With the OpenJ9 non-generational collectors `optavgpause` and `optthruput`, a NPE was happening in the notification listener which caused some code in it to not be executed. Namely, setting the live and max data size was being effectively skipped for these collectors.

Now we check for `null` and only the allocated bytes counter is skipped when no allocation pool name is set.