Reduce flakiness of JvmGcMetricsTest#sizeMetricsNotSetToZero
This sometimes fails if the notifications for gc aren't processed by the listener fast enough. A crude solution is sleeping for an arbitrary amount of time. It would be better to wait until all notifications are processed, but it may be fragile to assume how many notifications are expected, and it depends on the configured garbage collector. The test may not catch the failure it intends if the 100 millisecond sleep is not long enough to capture all notifications generated from a garbage collection. Asserting against a subset of notifications may miss the issue. Hopefully 100 ms is a reasonable amount of time in environments the tests are run for the notification listener to receive all notifications post-gc.