Exclude netty-codec-http from statsd module
We do not use HTTP in the statsd module, but netty-codec-http is brought in as a transitive dependency from reactor-netty-core. Our usage of reactor-netty-core works without this dependency. To reduce the amount of false positive reports we have to deal with, and to reduce the produced JAR size, we can exclude this unneeded dependency.

Resolves gh-2929