Update system.cpu.usage description (#2969)
The behavior of the JVM in the latest patch updates has changed since this metric was introduced. This provides a better description given the latest behavior when running inside a container. The system CPU usage is bounded by the resource restrictions set for the container, rather than the machine on which the container runs.