micrometer-api module split off from micrometer-core
This moves a portion of existing classes in micrometer-core to a new micrometer-api module. micrometer-core depends on the API module and now micrometer-core contains only the instrumentation that we maintain. Projects that want to write their own instrumentation now only need to depend on the micrometer-api module. This will save some kilobytes in the JAR size for them. End users may continue to depend on micrometer-core to get the instrumentation we provide in addition to the API. Package names are changed, which will require users to update their imports. We could not keep the same package name without causing split packages across micrometer-api and micrometer-core.