Native image reflection config for HdrHistogram usage
Our client-side percentiles use HdrHistogram as an implementation detail. That library uses some reflection that fails at runtime in a native image without this reflection config.
Ideally, the HdrHistogram library would work in native images out-of-the-box by, for example, having this reflection config in its published JAR. In the meantime, we will add the necessary config to Micrometer to provide a better experience for our users compiling to native images and using client-side percentiles.
With this, users should not need to add any configuration or hints themselves to use Micrometer in their native image applications.

Resolves gh-3002