Remove unnecessary methods from core Tag, avoid "common" term
The `compareTo` method should not be needed due to the inherited `compareTo` from the parent. Arrays and collections of core `Tag` can still be sorted using the inherited `compareTo`.

The `toCommon` method should also not be needed because a core `Tag` already is a `Tag` from commons. The inverse is not always true, which is why the other method `fromCommon` could be useful.

However, we want to avoid the term `common` when talking about tags unless we mean the already existing Micrometer concept of a common tag - that is, a tag applied to all meters by default. The `fromCommon` method has been renamed to `of` which is in line with existing API here.

A slight improvement to its implementation, it does not allocate a new object if the passed `Tag` already is an `instanceof` a core module `Tag`.