Use more specific Comparable type
It's not necessary to use `Object` to be able to compare either type `Tag` since the common module `Tag` is the base type, and therefore can always be used to compare another instance of its own type or the more specific core module `Tag`. In testing, there was no problem with this being used to sort Arrays and Collections of either type.