Improve consistency with unified code format/imports (#3156)
Configures the spring-javaformat plugin to enforce and apply code formatting as part of the build. Applies the formatting to existing code. Also adds to editorconfig configuration for import order and updates all imports. There is no generic support for Java import order configuration in editorconfig, so an IntelliJ specific configuration is used.