Config for OTLP Resource attributes (#3159)
* Config for OTLP Resource attributes

Adds a config method for attributes that are set on the Resource shared by all metrics published from the OtlpMeterRegistry. Of note, this allows configuring the `service.name` to identify your service, but is generic enough to allow configuration of arbitrary attributes.

The use cases for this largely overlaps with how we expect common tags are used, and it is a shame I could not come up with a way to better connect these now. There is precedent for this approach in e.g. the StackdriverMeterRegistry. But it potentially creates inconsistency or duplicated work when using multiple registries.

* Load resource attributes from config/env

Since there is a specified environment variable to set resource attributes, we can support that. And since it defines a format for key-value pairs, we can use that for loading the map from the `get` method.

* Make parsing more robust and tested

* Polish JavaDoc