Make descriptions static where they are not (#3201)
Since the description of the Meters can be different per Meter, we can run into situations when the description is dynamic/different for different Meter.Ids that share the same name but differ in tags. Some backends do not support different description per time series, so whichever is registered first will be used, and it can be confusing as the description is specific to one set of tag values.