Add a LongTaskTimer for every Observation (#3215)
Always create an additional LongTaskTimer (on top of the Timer) so that the active task can be observed too on-the-fly.

There is a caveat/issue with this: since the KeyValuesProvider is used at stop, none of those tags will be added to the LongTaskTimer. Also, the tags that are added to the Observation after start won't be added to the LongTaskTimer.